requirements-test.txt
coverage == 7.10.*
pre-commit == 4.3.*
pytest == 8.4.*
pytest-cov == 7.0.*
pytest-django == 4.11.*
tox == 4.30.*
